feat: 추천제품 페이지 하단 기획전 배너 색상 적용 & 링크 추가
 - 추천제품 페이지 하단 기획전 영역 Exhibitions Component로 분리
